Move python3 libraries to tests directory
They are used only for testing purposes.
